+++ CHANGES.md
@@ -2,6 +2,7 @@
 ### Improvements
  * Added bounds checking and suggestions for numbers in commands.
  * Added bounds checking and code suggestions for minecraft:time arguments in commands
+ * The schema context menu has been tidied up.
 
 
 ## 0.6.0-alpha